Make entity data mutable by default for simplicity
