+++ src/main/java/org/fountainmc/api/entity/data/EntityData.java
@@ -9,13 +9,17 @@ import org.fountainmc.api.NonnullByDefault;
 import org.fountainmc.api.Server;
 import org.fountainmc.api.entity.EntityType;
 
+import static com.google.common.base.Preconditions.*;
+
 /**
  * The data for an entity, which isn't necessarily spawned in a world.
- * <p>Notes:
- * <ul>
- * <li>this doesn't include the entity's location as that only makes sense for an entity that is in the world.</li>
- * <li>this doesn't include information about the riding entity because nbt doesn't contain that</li>
- * </ul>
+ * <p>
+ *     Notes:
+ *     <ul>
+ *         <li>this doesn't include the entity's location as that only makes sense for an entity that is in the world.</li>
+ *         <li>this doesn't include information about the riding entity because nbt doesn't contain that</li>
+ *         <li>you can't modify the passenger because it'd be extremely difficult and buggy to implement</li>
+ *     </ul>
  * </p>
  */
 @NonnullByDefault
@@ -36,11 +40,25 @@ public interface EntityData {
     float getPitch();
 
     /**
+     * Set the Pitch of the Entity.
+     *
+     * @param pitch Pitch to set the Entity to
+     */
+    void setPitch(float pitch);
+
+    /**
      * Get the Yaw of the Entity.
      *
      * @return the yaw
      */
     float getYaw();
+
+    /**
+     * Set the Yaw of the Entity.
+     *
+     * @param yaw Yaw to set the Entity to
+     */
+    void setYaw(float yaw);
 
     /**
      * Get the primary passenger riding on top of this entity
@@ -64,11 +82,15 @@ public interface EntityData {
 
     int getTicksOnFire();
 
+    void setTicksOnFire(int ticksOnFire);
+
     default boolean isOnFire() {
         return getTicksOnFire() >= 0;
     }
 
     boolean isImmuneToFire();
+
+    void setImmuneToFire(boolean immune);
 
     /**
      * The type of entity this data is intended for
@@ -78,4 +100,16 @@ public interface EntityData {
     @Nonnull
     EntityType<?> getEntityType();
 
+    /**
+     * Copy all of the given data to this object, making them effectively equal.
+     *
+     * @param data the data to copy from
+     */
+    default void copyDataFrom(EntityData data) {
+        checkNotNull(data, "Null data");
+        this.setPitch(data.getPitch());
+        this.setYaw(data.getYaw());
+        this.setTicksOnFire(data.getTicksOnFire());
+        this.setImmuneToFire(data.isImmuneToFire());
+    }
 }
