+++ src/main/java/org/fountainmc/api/entity/data/PlayerData.java
@@ -1,11 +1,12 @@
 package org.fountainmc.api.entity.data;
 
 import java.util.UUID;
-
 import javax.annotation.Nonnegative;
 
 import org.fountainmc.api.GameMode;
 import org.fountainmc.api.NonnullByDefault;
+
+import static com.google.common.base.Preconditions.*;
 
 /**
  * The data of a player.
@@ -34,14 +35,21 @@ public interface PlayerData extends LivingEntityData {
      */
     GameMode getGameMode();
 
+    void setGameMode(GameMode gameMode);
+
     boolean canFly();
+
+    boolean setCanFly(boolean canFly);
 
     @Nonnegative
     float getPercentageToNextExperienceLevel();
 
+    void setPercentageToNextExperienceLevel(@Nonnegative float percentage);
+
     @Nonnegative
     int getExperienceLevel();
 
+    void setExperienceLevel(@Nonnegative int experienceLevel);
 
     /**
      * Calculate and return the player's total experience
@@ -60,4 +68,42 @@ public interface PlayerData extends LivingEntityData {
         return total;
     }
 
+    default void setTotalExperience(long total) {
+        checkArgument(total >= 0, "Negative experience %s", total);
+        int level = 0;
+        while (true) {
+            long expAtLevel = getServer().getExpAtLevel(level);
+            if (expAtLevel <= total) {
+                total -= expAtLevel;
+                level++;
+            } else {
+                setPercentageToNextExperienceLevel((float) ((double) total / (double) expAtLevel));
+                break;
+            }
+        }
+        setExperienceLevel(level);
+    }
+
+    default void giveExp(@Nonnegative long amount) {
+        checkArgument(amount >= 0, "Can't give negative exp %s! Use takeExp(long) if you want that!", amount);
+        setTotalExperience(getTotalExperience() + amount);
+    }
+
+    default void takeExp(@Nonnegative long amount) {
+        checkArgument(amount >= 0, "Can't give negative exp %s! Use takeExp(long) if you want that!", amount);
+        long exp = getTotalExperience();
+        checkState(exp >= amount, "Can't take away %s exp because the player only has %s exp.", amount, exp);
+        setTotalExperience(exp - amount);
+    }
+
+    @Override
+    default void copyDataFrom(EntityData data) {
+        LivingEntityData.super.copyDataFrom(data);
+        if (data instanceof PlayerData) {
+            this.setGameMode(((PlayerData) data).getGameMode());
+            this.setCanFly(((PlayerData) data).canFly());
+            this.setPercentageToNextExperienceLevel(((PlayerData) data).getPercentageToNextExperienceLevel());
+            this.setExperienceLevel(((PlayerData) data).getExperienceLevel());
+        }
+    }
 }
