+++ src/main/java/org/fountainmc/api/entity/data/hanging/ItemFrameData.java
@@ -3,6 +3,7 @@ package org.fountainmc.api.entity.data.hanging;
 import javax.annotation.Nullable;
 
 import org.fountainmc.api.Material;
+import org.fountainmc.api.entity.data.EntityData;
 
 public interface ItemFrameData extends HangingEntityData {
 
@@ -23,4 +24,19 @@ public interface ItemFrameData extends HangingEntityData {
     @Nullable
     Material getMaterial();
 
+    /**
+     * Set the Material of the Item in the ItemFrame.
+     * <p>Setting it to null empties the item frame.</p>
+     *
+     * @param material the Material to set the Item in the ItemFrame to
+     */
+    void setMaterial(@Nullable Material material);
+
+    @Override
+    default void copyDataFrom(EntityData data) {
+        HangingEntityData.super.copyDataFrom(data);
+        if (data instanceof ItemFrameData) {
+            this.setMaterial(((ItemFrameData) data).getMaterial());
+        }
+    }
 }
