+++ src/main/java/org/fountainmc/api/entity/data/hanging/LeashKnotData.java
@@ -3,11 +3,17 @@ package org.fountainmc.api.entity.data.hanging;
 import javax.annotation.Nullable;
 
 import org.fountainmc.api.entity.data.EntityData;
-import org.fountainmc.api.entity.data.MutableEntityData;
 
 public interface LeashKnotData extends HangingEntityData {
 
     @Nullable
     EntityData getLeashedEntity();
 
+    @Override
+    default void copyDataFrom(EntityData data) {
+        HangingEntityData.super.copyDataFrom(data);
+        if (data instanceof HangingEntityData) {
+            // if we ever get anything to copy add it here
+        }
+    }
 }
